backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "teofils1", email = "teodor.marinescu03@stud.fils.upb.ro" }
]
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-model-utils>=5.0.0",
    "django-storages[s3]>=1.14.4",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.27.2",
    "psycopg[binary]>=3.2.9",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=7.1.0",
    "structlog>=24.4.0",
    "web3>=7.13.0",
    "weasyprint>=62.3",
    "celery>=5.4.0",
    "django-celery-beat>=2.7.0",
]

 

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.12.8",
]

[tool.uv]
dev-dependencies = [
    "ruff",
]

# Ruff configuration can also go here instead of ruff.toml
[tool.ruff]
line-length = 88
target-version = "py312"

 
[tool.hatch.build.targets.wheel]
packages = ["src/api", "src/supplychain"]

[tool.hatch.build.targets.sdist]
include = ["src/api", "src/supplychain", "README.md"]

